extracted EditCooperationPollForm
